backend/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3
django-prometheus==2.3
psycopg2-binary==2.9.9
celery[redis]==5.3
redis==5.0
channels==4.0
daphne==4.0
channels-redis==4.1
requests==2.31
python-dotenv==1.0
pytest==7.4
pytest-django==4.5
pytest-asyncio==0.21
factory-boy==3.3
coverage==7.3
